Fix traceSelected in explore-simple
Bug: 447094435

Fix traceSelected function in explore-simple. This function was
unconditionally using several uninitialized values, such as plot-simple,
or commits list. Since commits details tab was deprecated, we can remove
some of the code.

Proposed expected behavior: user selects a point on a trace, re-renders
the plot (e.g. by adding another test) - the selection should(?)
persist, and the plot should not crash/freeze. Note that explore-multi
simply hides tooltip on replot - perhaps explore-simple should do the
same?

Not sure what are the requirements for plotting when a point is
selected. This function was forgotten during migration to
plot-google-charts - as hinted by TODO(b/362831653).

Also, this function used to highlight whole trace within plot-simple.
Currently, we are just highlighting a legend entry in plot-google-chart.
This might be a missing feature.

Call to the /cid endpoint proved to be useless - but I am not sure if we
want this information to be displayed somewhere.

Looks like values produced by traceSelected are overwritten by other
functions - need to verify that. For example, ExploreSimpleSk.testPath
is overwritten when running addFromQueryOrFormula.

Change-Id: I5954cc709d30cdc01aa82a64fbba6cb02a5dd3e1
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1064396
Reviewed-by: Tony Seaward <seawardt@google.com>
Commit-Queue: Marcin Mordecki <mordeckimarcin@google.com>