Remove more plot-simple references from explore-simple
Bug: 408212152

Removing most of leftover plot-simple references. Leaving
show_google_plot references for now, will remove later.

Explore-demo is in quite a bad shape. I migrated it so it doesn't break
when we remove plot-simple, but it needs some polishing. This will be
done in https://buganizer.corp.google.com/issues/450184956.

Also, Gold properly identified a difference: the missing "New Query"
button was only visible for a split-second. If we use plot-google-chart,
we open the dialog automatically, so there's no need for a user action.

Change-Id: I3371d684ddebb7d0152fe274b15c5f9256321d87
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1068996
Commit-Queue: Marcin Mordecki <mordeckimarcin@google.com>
Reviewed-by: Sergei Rudenkov <sergeirudenkov@google.com>