[task scheduler] Skip Swarming pool task templates
We're getting conflicts with the task template for a chrome pool for
some of our test machines.

If we find that this is a problem for whatever reason (eg. we want to
start using task templates for Skia pools), we can make this
configurable per-pool using a flag, or possibly make it part of the task
spec to configure it on a per-task basis.

Bug: b/443292328
Change-Id: Ia76278e724a171369262f375b7def0f1dad8f9d0
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1069796
Commit-Queue: Eric Boren <borenet@google.com>
Reviewed-by: Kaylee Lubick <kjlubick@google.com>