[CBB] Fix nil pointer error in release detection workflow
Previously, httpClient was initialized in CbbNewReleaseDetectorWorkflow
and then used in GetChromeReleasesInfoActivity. Temporal may chose to
run the workflow and the activity on different workers, so the variable
might be uninitialized while running the activity. This CL moves where
the variable is initialized to avoid this problem.

Bug: b/433796487
Change-Id: I1e365fa020996b09cdc66734a324b14dde04bbca
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1070166
Commit-Queue: John Chen <zhanliang@google.com>
Reviewed-by: Farid (Mojtaba) Faridzad <faridzad@google.com>