Added an input box to the user-issue-sk component.
Re-added the input box, which allows users to associate an existing bug with a datapoint. The logic now first checks if the input bug exists. If it does, a link to the bug is shown; else, a new issue is created, and then the bug link is shown.

Bug: b/424207630


Change-Id: I5537ab62fe5375df6aa04e9a6569c08ea9a0ec85
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1073294
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Maggie Dong <jiaxindong@google.com>