Fix duplicate IDs in anomalies-table-sk
This change introduces a unique ID for each instance of
`AnomaliesTableSk` and appends it to the IDs of internal elements
(checkboxes, buttons, tables). This prevents ID collisions when multiple
instances of the component are present on the same page, ensuring that
labels and query selectors target the correct elements.

perfui (fix-anomalies-ids)

Change-Id: If5c370697c51b6e44eed19eec59a368828952056
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1099873
Commit-Queue: Tony Seaward <seawardt@google.com>
Reviewed-by: Farid (Mojtaba) Faridzad <faridzad@google.com>