Perf UI: Add configurable header image and move images
Adds a new configuration option 'header_image_url' to InstanceConfig.
Updates the frontend to use this URL for the header logo if provided.
Defaults to the Chrome logo if not specified.
Adds the V8 logo to the distribution and configures v8-internal to use it.
Moves chrome-logo.svg and line-chart.svg from am/images to perf/images.
Updates BUILD files to reflect the new location.

skiaui (feature-instance-name)

Change-Id: Ia402a1012e7274c7d2be999b2e730bc2b30ac92d
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1100558
Reviewed-by: Farid (Mojtaba) Faridzad <faridzad@google.com>
Commit-Queue: Tony Seaward <seawardt@google.com>