[bazel] Fix directory ownership in docsyserver image
This is a weird problem with a kind of crazy solution.

There is a quirk in the way that pkg_tar interacts with container_image
which can result in problems with directory ownership. The
skia_app_container macro creates a separate pkg_tar for each file added
to the image. Because the resulting tar files have no knowledge of the
directory layout and ownership of the previous layer, each tar file will
build a path to the specified file, with all of the containing
directories having the same owner as the one specified for the file.
The tar files are merged to become a layer on top of the previous layer,
which causes any existing directories from the previous layer to be
overwritten by the directories from the tar files, including their
ownership. This creates complications when we add files with a
particular owner (the default being root) to directories with other
owners (eg. /home/skia), causing the user's home directory to be owned
by root, or causing "/home" to be owned by "skia", etc.

Our solution to this is to create a fixup layers after the pkg_tar layer
which overlay the correct ownership onto each directory. In order to do
this, we need the caller to specify which directories should be owned by
which user, and then we build a tree containing all of the paths that
were given, propagating the correct ownership to each directory. Then,
we find the subtrees which have the same owner, sort them by depth, and
create layers from them to ensure that each set of directories has the
correct owner.

Bug: b/458681039
Change-Id: I92e7410389f5bf26c0f80de39da26750fcac83f9
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1109736
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Eric Boren <borenet@google.com>