Fix: update Redis instance names in chrome-internal-ng and v8-internal-ng perf configs
Currently new instances are crash-looping because there is no Redis.

Update configs to share the cache with the main instance. This allows us
to avoid extra steps and cost of setting up separate Redis instances,
and it should be safe to share this cache with main instance (because
for now, we don't store anomalies in the Redis cache). For example, our
LTS instance shares the cache with non-LTS instance.

Bug: 462378384
Change-Id: I33a9778e93d0f13ff371c6172eb237254cfca0c8
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1111241
Reviewed-by: Sergei Rudenkov <sergeirudenkov@google.com>
Commit-Queue: Anri Sidorov <ansid@google.com>