anomalies playground. backend
This is playground for us as developers to debug why anomaly was detected or not. (We can get the dataframe from db and run different algorithms against it)" We can easily copy real data from the graph we're interested in from the devtools (document.querySelector('explore-simple-sk')._dataframe.traceset). When it's time to expose it to users, we can add a button to do exactly that.

Demo: https://screencast.googleplex.com/cast/NTgyMjA3MjQxOTc3ODU2MHwwY2I4NjA1ZC1lZQ

Bug: 463952896
Change-Id: I565871a83324592f1ff559d7f55052ed93f90283
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1111257
Reviewed-by: Anri Sidorov <ansid@google.com>
Commit-Queue: Sergei Rudenkov <sergeirudenkov@google.com>