anomalies playground. frontend
This is playground for us as developers to debug why anomaly was detected or not. (We can get the dataframe from db and run different algorithms against it)" We can easily copy real data from the graph we're interested in from the devtools (document.querySelector('explore-simple-sk')._dataframe.traceset). When it's time to expose it to users, we can add a button to do exactly that.

Bug: 463952896
Change-Id: I89752acf6a1c30ea2db7a7a2904e9a7a7ed5c532
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1111960
Reviewed-by: Anri Sidorov <ansid@google.com>
Commit-Queue: Sergei Rudenkov <sergeirudenkov@google.com>