[bazel] Replace extract_files_from_skia_wasm_container
The existing implementation uses rules_docker, which we need to move awa
from. It was fairly straightforward to write a script to extract files
from the image tarball(s).

Note: this drops the existing functionality where we avoid pulling the specified image (and pull "@empty_container" instead) when the resulting files are not going to be used. As I understand it, this should not be a problem because when the flag is set we don't depend on the output files anyway, so Bazel shouldn't attempt to build the target.
Bug: b/458681039
Change-Id: Ib02727086e96d712f8e480009ec5e4e0ffaaf422
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1115076
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Eric Boren <borenet@google.com>