[bazel] Make temporal build using Bazel, not Go-inside-Docker
We have to move off of rules_docker, which means we can no longer use
the container_run_and_extract rule. Setting up the temporal repos as Go
dependencies and building the binaries using Bazel is a cleaner way to
do this.

I tried to keep the same versions of the dependencies, but I was forced
to downgrade temporal-ui from 2.27.3 to 2.26.2 to keep the dependency
graph consistent. If this is a problem, we'll need to update all of the
temporal repos together.

Note that some files have different destination paths within the Docker
images than the current rules would produce. This was intentional; the
builds for these images have been broken for months and I did my best to
mimic the images which resulted from the most recent successful builds.

Bug: b/458681039
Change-Id: Ic554e543d2dfa35eaec1b50b72d4dd423d267f07
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1117437
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Eric Boren <borenet@google.com>