feat: Add discrete axis feature flag in plot-google-chart-sk
This change introduces a new feature flag  to toggle the discrete axis feature in the plot-google-chart-sk component.

When the  URL parameter is set to , the component will use a discrete axis for the chart. This allows data points to be evenly spaced on the X axis, making it easier to judge the distribution of a metric, including when it's bimodal or has outliers, similar to Chromeperf's graphs.

Currently, Skia Perf scales the X axis by commit number or time, which can distort the visual representation of data distribution due to varying commit/build rates. This new option provides a more consistent view.

This change is made to ensure that the new feature will not affect anything unless the flag is set.

Bug: 418808644
Change-Id: I166df012da1d95ae7d96092bb3aa5b49f0bbaf10
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1117924
Reviewed-by: Tony Seaward <seawardt@google.com>
Commit-Queue: Eduardo Yap <eduardoyap@google.com>