Fix: Force horizontal labels in discrete axis [perf]
When using a discrete axis, the labels would show up as '...' when there were few points, but display fine with many points.

This was likely due to Google Charts' default minTextSpacing, which added padding between labels. When the number of points was small, the allocated space for each label was smaller, and the padding caused the text to be truncated.

This change forces the labels to be horizontal and removes the default spacing to allow the labels to fit.

Bug: 418808644
Change-Id: I7f3d494be15634c7106d5daac990eda8dafba4ce
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1120556
Reviewed-by: Tony Seaward <seawardt@google.com>
Commit-Queue: Eduardo Yap <eduardoyap@google.com>
Reviewed-by: Sergei Rudenkov <sergeirudenkov@google.com>