Add support for loading graph configurations from JSON.
Bug: b/462436251

This change introduces a new "JSON" tab in the query dialog, allowing users to paste a JSON object containing graph definitions (queries, formulas, keys). The component can also load this JSON from a 'json' URL parameter. The clearBrowserURL function is updated to clear all search parameters.

It could be tested with this example of json

```
{
  "graphs": [
    {
      "queries": ["benchmark=jetstream2&bot=mac-m1-pro-perf&stat=value&test=3d-cube-SP"],
      "formulas": [],
      "keys": ""
    }
  ]
}
```

Change-Id: I16fcf7641890a71af76d4d38663dcf3f39db0899
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1111238
Reviewed-by: Sergei Rudenkov <sergeirudenkov@google.com>