Add keyboard shortcuts to the triage page
Bug: b/445051442

Introduce keyboard shortcuts for common actions when the triage dialog is
open, significantly improving the efficiency of the triage workflow.

The following shortcuts are now available:

* **p**: Mark the cluster as positive.
* **n**: Mark the cluster as negative.
* **g**: Open the report for the selected cluster.
* **?**: Show the help dialog with all registered shortcuts.

This change also refactors methods in `cluster-summary2-sk` from private
to public to allow `triage-page-sk` to invoke them for the shortcuts. A
new `keyboard-shortcuts-help-sk` element and a help button have been
added to the triage page to display the available shortcuts.

Change-Id: I5f6e580720bfc98adafe4173eb37701c6bdb4a65
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1112056
Reviewed-by: Sergei Rudenkov <sergeirudenkov@google.com>